storage: minor fixes — const dedup, cast direction, emulator gate, schema, digests
- service.rs: remove private IDEMPOTENCY_KEY_HEADER const (duplicated from
  wyrd-spec); import the canonical constant from wyrd_spec::storage instead

- sweeper.rs: change DEFAULT_BATCH_SIZE and DEFAULT_IDEMPOTENCY_BATCH_SIZE from
  i64 to u64 (eliminating the roundtrip as u64 cast in parse_clamped_i64); add
  as i64 casts at the struct field and Default impl sites

- migrations/storage_idempotency.sql: add comment explaining the split-string
  regex technique so future maintainers know not to collapse the concatenation

- factory/gcs.rs + factory/azure.rs: gate build_emulator_signer behind
  #[cfg(any(test, feature = "emulator"))] to prevent production binaries from
  calling emulator constructors; add emulator feature to wyrd-storage/Cargo.toml

- docker-compose.yml: add SHA256 digests for minio/minio, fsouza/fake-gcs-server,
  and azurite; remove pull-instructions comment (digests are self-documenting);
  minio/mc pinned tag was not present on registry, left unchanged

- upload.rs: fix stale GCS doc comment on expected_sha256 (was server-computed,
  now client-declared); add LocalBlobUploadResponse type so the local blob PUT
  endpoint is covered by codegen:check and the JSON schema catalog

- routes.rs: return LocalBlobUploadResponse instead of serde_json::Value for
  the local blob handler

- gen_schemas.rs + schema goldens: register and regenerate LocalBlobUploadResponse

+++ crates/wyrd/wyrd-storage/src/sweeper.rs
@@ -23,11 +23,11 @@ pub const SWEEPER_LEADER_LOCK_KEY: i64 = 0x5759_7264_5374_6f72_i64;
 /// Default interval between sweeper ticks.
 pub const DEFAULT_TICK: Duration = Duration::from_mins(1);
 /// Default maximum multipart rows processed per tick.
-pub const DEFAULT_BATCH_SIZE: i64 = 100;
+pub const DEFAULT_BATCH_SIZE: u64 = 100;
 /// Default age after which `initiating` rows are treated as orphaned.
 pub const DEFAULT_INIT_GRACE: Duration = Duration::from_secs(30);
 /// Default maximum idempotency rows deleted per tick.
-pub const DEFAULT_IDEMPOTENCY_BATCH_SIZE: i64 = 500;
+pub const DEFAULT_IDEMPOTENCY_BATCH_SIZE: u64 = 500;
 
 const ENV_ENABLED: &str = "WYRD_STORAGE_SWEEPER_ENABLED";
 const ENV_TICK_SECS: &str = "WYRD_STORAGE_SWEEPER_TICK_SECS";
@@ -66,7 +66,7 @@ impl SweeperConfig {
             ),
             batch_size: parse_clamped_i64(
                 ENV_BATCH_SIZE,
-                DEFAULT_BATCH_SIZE as u64,
+                DEFAULT_BATCH_SIZE,
                 1,
                 1000,
             )?,
@@ -77,7 +77,7 @@ impl SweeperConfig {
             ),
             idempotency_batch_size: parse_clamped_i64(
                 ENV_IDEMPOTENCY_BATCH_SIZE,
-                DEFAULT_IDEMPOTENCY_BATCH_SIZE as u64,
+                DEFAULT_IDEMPOTENCY_BATCH_SIZE,
                 1,
                 5000,
             )?,
@@ -90,9 +90,9 @@ impl Default for SweeperConfig {
         Self {
             enabled: true,
             tick: DEFAULT_TICK,
-            batch_size: DEFAULT_BATCH_SIZE,
+            batch_size: DEFAULT_BATCH_SIZE as i64,
             init_grace: DEFAULT_INIT_GRACE,
-            idempotency_batch_size: DEFAULT_IDEMPOTENCY_BATCH_SIZE,
+            idempotency_batch_size: DEFAULT_IDEMPOTENCY_BATCH_SIZE as i64,
         }
     }
 }
@@ -407,9 +407,9 @@ mod tests {
 
             assert!(cfg.enabled);
             assert_eq!(cfg.tick, DEFAULT_TICK);
-            assert_eq!(cfg.batch_size, DEFAULT_BATCH_SIZE);
+            assert_eq!(cfg.batch_size, DEFAULT_BATCH_SIZE as i64);
             assert_eq!(cfg.init_grace, DEFAULT_INIT_GRACE);
-            assert_eq!(cfg.idempotency_batch_size, DEFAULT_IDEMPOTENCY_BATCH_SIZE);
+            assert_eq!(cfg.idempotency_batch_size, DEFAULT_IDEMPOTENCY_BATCH_SIZE as i64);
         });
     }
